Cleaned solution
