+++ TagsCloudContainer/TagsCloudContainer.Interfaces/CloudWord.cs
@@ -1,3 +1,3 @@
-namespace TagsCloudContainer;
+namespace TagsCloudContainer.Interfaces;
 
 public record CloudWord(string Text, int Occurrences);
