+++ TagsCloudContainer/TagsCloudContainer.Interfaces/DrawerSettings.cs
@@ -1,0 +1,5 @@
+﻿namespace TagsCloudContainer.Interfaces;
+
+public abstract class DrawerSettings
+{
+}
