+++ TagsCloudContainer/TagsCloudContainer.Interfaces/IDrawer.cs
@@ -1,4 +1,4 @@
-﻿namespace TagsCloudContainer;
+﻿namespace TagsCloudContainer.Interfaces;
 
 public interface IDrawer
 {
