+++ TagsCloudContainer/TagsCloudContainer.Interfaces/IDrawerFactory.cs
@@ -1,7 +1,6 @@
 ﻿using System.Drawing;
-using TagsCloudContainer.Interfaces;
 
-namespace TagsCloudContainer;
+namespace TagsCloudContainer.Interfaces;
 
 public interface IDrawerFactory
 {
