+++ TagsCloudContainer/TagsCloudContainer.Interfaces/IWordsStream.cs
@@ -1,0 +1,5 @@
+﻿namespace TagsCloudContainer.Interfaces;
+
+public interface IWordsStream
+{
+}
