+++ TagsCloudContainer/TagsCloudContainer.Interfaces/TagsCloudContainer.Interfaces.csproj
@@ -6,4 +6,8 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
+    <ItemGroup>
+        <PackageReference Include="System.Drawing.Common" Version="7.0.0"/>
+    </ItemGroup>
+
 </Project>
