+++ TagsCloudContainer/TagsCloudContainer/ClassicDrawerSettings.cs
@@ -1,4 +1,5 @@
 ﻿using System.Drawing;
+using TagsCloudContainer.Interfaces;
 
 namespace TagsCloudContainer;
