+++ TagsCloudContainer/TagsCloudContainer/DeepMorphyFunnyWordsSelector.cs
@@ -1,4 +1,5 @@
 using DeepMorphy;
+using TagsCloudContainer.Interfaces;
 
 namespace TagsCloudContainer;
