--- TagsCloudContainer/TagsCloudContainer/DrawerSettings.cs
@@ -1,5 +1,0 @@
-﻿namespace TagsCloudContainer;
-
-public abstract class DrawerSettings
-{
-}
