--- TagsCloudContainer/TagsCloudContainer/IWordsStream.cs
@@ -1,5 +1,0 @@
-﻿namespace TagsCloudContainer;
-
-public interface IWordsStream
-{
-}
